// src/types/landingpage/LandingpageTypes.ts
type ProductsSliderTypes = {
    type:boolean,
    img: string;
    name: string;
    link: string;
};

type UserReviewTypes = {
    img: string;
    title: string;
    subtitle: string;
    review: string;
};

type ListFeatureTypes = {
    icon: any;
    title: string;
    subtitle: string;
};

type DemosMegaMenuTypes = {
    img: string;
    name: string;
    link: string;
};

type AppsMegaMenuTypes = {
    img: string;
    name: string;
    link: string;
};

export type { ProductsSliderTypes, UserReviewTypes, ListFeatureTypes, DemosMegaMenuTypes, AppsMegaMenuTypes };